add video player & separate components
